v1 Delta 1.10
Another try to update state from SongList component: It seems like the connect component or the createStore component is not working properly. I did troubleshoot in many ways by console logging out different values, almost everything is fine.

Look ./src/components/SongList.js , here selectSong component (action creator) is imported and at line 38 that is connected with redux using the connect component's second argument (aka mapDispatchToProps) and the class component SongList is invoked. When the props in the component is console logged, the dispatch function is missing, that makes me doubt that the connect function or createStore or the Provider is not working properly. Probably there's any new update that changed redux behaviours or something else. If anybody can fix this, please do a pull request.