Version 1 Release
This is finishment of the app. This app is a basic text and lab project for understanding react-redux. Unfortunately initially I used the deprecated redux methods which were pretty famous couple of years back. Then on commit 'v1 Delta 1.10' as I could not be able to update state data from a component, I read some documentations and understood my mistakes. Then I recreated the logics using the latest practice of redux as reduxjs. This app was a struggle, but I did learnt something.

Right now the app contains a list of songs and when user clicks on the select buttons below any song, that song gets selected and details of that song is rendered on a card beside the list.